Drop the Tcs structure
This structure is mostly useless, except in trivial testing cases. We
therefore retain it in the enclave creation test. But it is no logner
public.
